Refactor the stats module
