+++ app/core/task/stats/ErrorStats.py
@@ -1,0 +1,2 @@
+class ErrorStats:
+    is_error = False
