+++ app/core/task/stats/ValidateToolStats.py
@@ -1,0 +1,51 @@
+from typing import Any
+from typing import Dict
+
+from app.core.task.stats.PatchStats import PatchStats
+from app.core.task.stats.ToolStats import ToolStats
+
+
+class ValidateToolStats(ToolStats):
+    patch_stats: PatchStats
+    bug_info: Dict[str, Any]
+    config_info: Dict[str, Any]
+
+    def __init__(self):
+        self.patch_stats = PatchStats()
+        super(ValidateToolStats, self).__init__()
+
+    def get_dict(self):
+        res = super(ValidateToolStats, self).get_dict()
+        res["details"]["space"] = self.patch_stats.get_dict()
+        if "info" not in res:
+            res["info"] = dict()
+        res["info"]["bug-info"] = self.bug_info
+        res["info"]["config-info"] = self.config_info
+        return res
+
+    def write(self, printer, prefix=""):
+        printer("{1} search space size: {0}\n".format(self.patch_stats.size, prefix))
+        printer(
+            "{1} count enumerations: {0}\n".format(
+                self.patch_stats.enumerations, prefix
+            )
+        )
+        printer(
+            "{1} count invalid patches: {0}\n".format(
+                self.patch_stats.non_compilable, prefix
+            )
+        )
+
+        printer(
+            "{1} count incorrect patches: {0}\n".format(
+                self.patch_stats.get_implausible(), prefix
+            )
+        )
+
+        printer(
+            "{1} count plausible patches: {0}\n".format(
+                self.patch_stats.plausible, prefix
+            )
+        )
+
+        super(ValidateToolStats, self).write(printer, prefix)
