+++ app/core/task/typing/TaskType.py
@@ -1,4 +1,6 @@
 from typing import Literal
 
 
-TaskType = Literal["fuzz", "repair", "analyze", "prepare"]
+TaskType = Literal[
+    "fuzz", "repair", "composite", "localize", "select", "analyze", "prepare"
+]
