+++ app/drivers/tools/AbstractTool.py
@@ -14,7 +14,7 @@ from app.core import definitions
 from app.core import emitter
 from app.core import utilities
 from app.core import values
-from app.core.task.stats import ToolStats
+from app.core.task.stats.ToolStats import ToolStats
 from app.core.task.TaskStatus import TaskStatus
 from app.core.task.typing.DirectoryInfo import DirectoryInfo
 from app.core.utilities import error_exit
