+++ app/drivers/tools/fuzz/AbstractFuzzTool.py
@@ -6,7 +6,7 @@ from os.path import join
 from app.core import definitions
 from app.core import utilities
 from app.core import values
-from app.core.task.stats import FuzzToolStats
+from app.core.task.stats.FuzzToolStats import FuzzToolStats
 from app.core.utilities import error_exit
 from app.drivers.tools.AbstractTool import AbstractTool
