+++ app/drivers/tools/select/AbstractSelectTool.py
@@ -8,7 +8,7 @@ from typing import List
 from app.core import container
 from app.core import definitions
 from app.core import utilities
-from app.core.task.stats import SelectToolStats
+from app.core.task.stats.SelectToolStats import SelectToolStats
 from app.drivers.tools.AbstractTool import AbstractTool
